% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{bmiThresholdDoc}
\alias{bmiThresholdDoc}
\title{Helper for consistent documentation of \code{bmiThreshold} argument}
\arguments{
\item{bmiThreshold}{Argument to indicate the thresholds for the obesity using
BMI measurements. It can be:
\itemize{
\item A single number, any BMI measurement above the threshold will be consider
as an obesity record.
\item A tibble with the columns \code{bmi_threshold}, \code{sex}, \code{age_min} and \code{age_max},
to use age and sex specific thresholds.
\item NULL the table \code{CohortRecipes::bmiThreshold} will be used.
}}
}
\description{
Helper for consistent documentation of \code{bmiThreshold} argument
}
\keyword{internal}
