% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRecipes-package.R
\docType{package}
\name{CohortRecipes-package}
\alias{CohortRecipes}
\alias{CohortRecipes-package}
\title{CohortRecipes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortRecipes/}
  \item Report bugs at \url{https://github.com/ohdsi/CohortRecipes/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
