% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameCohortDoc}
\alias{nameCohortDoc}
\title{Helper for consistent documentation of \code{name} argument in cohorts}
\arguments{
\item{name}{A character string with the name of the new cohort.}
}
\description{
Helper for consistent documentation of \code{name} argument in cohorts
}
\keyword{internal}
