% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity.R
\name{obesityCohort}
\alias{obesityCohort}
\title{Create the obesity cohort}
\usage{
obesityCohort(cdm)
}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object.}
}
\value{
The instantiated cohort
}
\description{
Create the obesity cohort
}
\examples{
library(CohortRecipes)

}
