% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity.R
\name{addObesity}
\alias{addObesity}
\title{Add obesity flag}
\usage{
addObesity(
  x,
  indexDate = "cohort_start_date",
  window = c(-Inf, 0),
  conceptSet = NULL,
  bmiThreshold = NULL,
  nameStyle = "obesity",
  name = tableName(x)
)
}
\arguments{
\item{x}{A \code{cdm_table} object, it mus contain \code{person_id} or \code{subject_id} as
columns.}

\item{indexDate}{A character string that points to a \code{Date} column in the \code{x}
table.}

\item{window}{Window to asses \code{bmi} in, it must be a vector of two numeric values \code{c(min, max)}. Window times refer to days since \code{indexDate}.}

\item{conceptSet}{It can either be a \if{html}{\out{<codelist>}}, <codelist_with_details> or <concept_set_expression> object. It must contain \code{bmi} as concepts. If \code{NULL} concepts will be retrieved using the OmopConcepts package.}

\item{bmiThreshold}{Argument to indicate the thresholds for the obesity using
BMI measurements. It can be:
\itemize{
\item A single number, any BMI measurement above the threshold will be consider
as an obesity record.
\item A tibble with the columns \code{bmi_threshold}, \code{sex}, \code{age_min} and \code{age_max},
to use age and sex specific thresholds.
\item NULL the table \code{CohortRecipes::bmiThreshold} will be used.
}}

\item{nameStyle}{A character string with the name of the new column.}

\item{name}{A character string with the name of the new table. If \code{NULL} a
temporary table will be created.}
}
\value{
The table \code{x} with a new column column indicating if subjects have
an obesity record in the window of interest.
}
\description{
Add obesity flag
}
\examples{
library(CohortRecipes)



}
