% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRecipes-package.R
\docType{package}
\name{CohortRecipes-package}
\alias{CohortRecipes}
\alias{CohortRecipes-package}
\title{CohortRecipes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
