% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity.R
\name{obesityCohort}
\alias{obesityCohort}
\title{Create the obesity cohort}
\usage{
obesityCohort(cdm, name, conceptSet = NULL, bmiThreshold = NULL)
}
\arguments{
\item{cdm}{A \verb{<cdm_reference>} object.}

\item{name}{A character string with the name of the new cohort.}

\item{conceptSet}{It can either be a \if{html}{\out{<codelist>}}, <codelist_with_details> or <concept_set_expression> object. It must contain \code{obesity}, \code{bmi} as concepts. If \code{NULL} concepts will be retrieved using the OmopConcepts package.}

\item{bmiThreshold}{Argument to indicate the thresholds for the obesity using
BMI measurements. It can be:
\itemize{
\item A single number, any BMI measurement above the threshold will be consider
as an obesity record.
\item A tibble with the columns \code{bmi_threshold}, \code{sex}, \code{age_min} and \code{age_max},
to use age and sex specific thresholds.
\item NULL the table \code{CohortRecipes::bmiThreshold} will be used.
}}
}
\value{
The cohort 'obesity' object.
}
\description{
Create the obesity cohort
}
\examples{
\donttest{
library(CohortRecipes)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

cdm$obesity <- obesityCohort(cdm = cdm, name = "obesity")

cdm$obesity

}

}
